Add TTL_fetch_tensor and some cpp samples to use it
TTL_fetch_tensor allows a tensor to be passed easily from the
host to the device. The host sends a ptr to the tensor structure
using standard opencl methods.

The Kernel then modes the tensor to __local memory using the
TTL_fetch_tensor method.

Examples showing using of this are in opencl/samples/cpp

A minor change included is to rename all the pipeline scheme
finish functions TTL_finish_buffering. This provides greater
consistancy

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>