Demonstrate passing a Tensor from Host to Kernel
Update the c and CPP samples to pass the Tensor from the Host to
the Kernel.  This demonstrates TTL structures and methods being
used on the Host and the Device side.

Passing of structures with pointers is only support in OpenCL 3.0
and only implemented in clang 17, for this reason a bug in clang
is used to allow the passing of a pointer.

For speed that validaty of a parameter passed to the Kernel is
cached. If a pointer is passed outside of a struct and then in a
struct the validaty of the pointer within the struct is not checked
but cached as OK.  The current code makes use of the bug to allow
the demonstration. With Clang 17 this workaround would not be
required.

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>