Add github actions to build and Deploy Doxygen Pages
Add a github action to build and deploy doxygen plages to
the gh-pages branch.

See online help for details of the build-and_deploy_doxygen.yml
file contents. The README.md has been changed to provide a link
to the online help.

This solution means only the tip of main is documented online,
which is a standard approach.

Doxygen is built and checked for every push, it is a requirement
for this to pass before landing is possible. When main is pushed
the updated Doxygen is deployed.

This patch deploys the Doxygen as is. The fixing up of the any
errors should be carried out in specific patches.

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>