Add several quality of life improvements for doxygen:
Strips the base directory from the path, so there is no github/workspace prefix in the file list.

Enables C syntax highlighting and more for the .cl files.

Removes the *.md and *.txt files from Doxygen. Although the introduction doc, this is probably
not wanted. Since doxygen grabs everything from the root folder (which is where the library is),
it currently grabs unwanted *.txt and *.md files.

Adds a nice theme, with small modifications, so it's nicer to work with on a desktop and mobile.