Add TTL_read_tensor and TTL_write_tensor
For both Tensors and SubTensors it makes a lot of sense
to have standardized and typed read and write functions.

The implmentation is as follows.

TTL_read_[type]_tensor(tensor, [value_for_write], x [,y [, z]])

The return or value is the same as the type of the tensor, at
present no limits checking is performed.

No method exists for a void_tensor - a tensor with a type of void
as this cannot be return.

All the samples and tests are updated to include this new method and
the test directory is updated with some addition changes due to
tensor typing

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>