Add CI runtime check and use available local memory
A small number of bugs/issues had crept into the code and
so a CI runtime of the C samples as been added to try to prevent
this.

The issues fixed are

Not using the actually available local memory to allocate buffers,
this is fixed by querying the amount of available memory before
complication.

The memory of the targets is now queried by the host side before
compilitation and based as a macro called LOCAL_MEMORY_SIZE.

A second macro for the samples with a host side written in Python
is to pass the printf specifier for the type to remove a warning
when the Tensor type is long or unsigned long.

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>