Add ability to record and replay lists of ASync copies
By linking togeather a multiple ASync copies into a single
user level transation the opportunity for gather scatter
type operations becomes possible.

This patch only allows this in the row axis, which may be
the only axis that makes sense due to the nature of the
underlying storage.

An income tensor can be gathered in any order, for example.

A 10 row global Tensor could be mapped to a 4 row local tensor
with the following mapping
Global Row 1 -> Local Row 1
Global Row 6,7 -> Local Row 2,3
Global Row 3 -> Local Row 4

The mapping is passed to the tiler, and from that point forwards
is invisible to the user.