Improve the Doxygen code to be post processed.
Although Doxygen tries to deal with Macro expansion, the level of
expansion in TTL defeats it.

Now TTL can produce a post-processed file - and by reprocessing this
file to produce the original files and then running doxygen on these
files, we can produce better Doxygen dealing with all of the types
generated.

The process is.
 1. Produce the single post-processor file.
 2. Split this file into the files indicated by the preprocessor
    # n "filename.h" n tags.

Run Doxygen on these regenerated files.

This does mean the Doxygen is lightly detached from the original source
but it is more useful as a documentation tool.

Signed-off-by: Chris Gearing <chris.gearing@mobileye.com>