Create a CPP implementation of TTL #37
The C implementation of TTL is really an object-orientated CPP
implementation written in C.

The outcome for the use is good, with the ability to work well with
Typed-Tensors, but the developer of TTL has to work with quite difficult
macros. The end user could also benefit from a CPP implementation.

Create a Prototype implementation.

At present a CPP (not opencl) version is implemented that makes Duplex
buffering work.