Create a CPP implementation of TTL #37
The C implementation of TTL is really an object-orientated CPP
implementation written in C.

The outcome for the user is good, with the ability to work well with
Typed-Tensors, but the developer of TTL has to work with quite difficult
macros. The end user could also benefit from a CPP implementation.

This patch adds a parallel C++ version with the same functionality of
the 'C' version. The idea being to phase out the 'C' version and move
to the CPP version.

This version does not yet build a CPP Kernel and run under opencl.